Rename LibraryRoomAvailability.tsx to FacilityRoomDetails.tsx for clarity
