+++ components/FacilityRoomDetails.tsx
